Update Axis Highlight
