+++ packages/x-charts/src/ChartsAxisHighlight/ChartsAxisHighlight.types.ts
@@ -1,0 +1,6 @@
+export type ChartsAxisHighlightType = 'none' | 'line' | 'band';
+
+export type ChartsAxisHighlightProps = {
+  x?: ChartsAxisHighlightType;
+  y?: ChartsAxisHighlightType;
+};
