+++ packages/x-charts/src/ChartsAxisHighlight/chartsAxisHighlightClasses.ts
@@ -1,0 +1,18 @@
+import generateUtilityClass from '@mui/utils/generateUtilityClass';
+import generateUtilityClasses from '@mui/utils/generateUtilityClasses';
+
+export interface ChartsAxisHighlightClasses {
+  /** Styles applied to the root element. */
+  root: string;
+}
+
+export type ChartsAxisHighlightClassKey = keyof ChartsAxisHighlightClasses;
+
+export function getAxisHighlightUtilityClass(slot: string) {
+  return generateUtilityClass('MuiChartsAxisHighlight', slot);
+}
+
+export const chartsAxisHighlightClasses: ChartsAxisHighlightClasses = generateUtilityClasses(
+  'MuiChartsAxisHighlight',
+  ['root'],
+);
