+++ packages/x-charts/src/ChartsAxisHighlight/index.ts
@@ -1,1 +1,4 @@
 export * from './ChartsAxisHighlight';
+export * from './chartsAxisHighlightClasses';
+export * from './ChartsAxisHighlight.types';
+export * from './ChartsAxisHighlightPath';
